+++ tests/schemas/test_yaml.py
@@ -20,7 +20,7 @@ FN = "yaml.cbz"
 TMP_DIR = get_tmp_dir(__file__)
 TEST_EXPORT_PATH = TMP_DIR / MetadataFormats.COMICBOX_CLI_YAML.value.filename
 YAML_PATH = TEST_METADATA_DIR / MetadataFormats.COMICBOX_CLI_YAML.value.filename
-YAML_NOTES = TEST_READ_NOTES + " urn:comicvine:issue:145269"
+YAML_NOTES = TEST_READ_NOTES + " urn:comicvine:145269"
 
 READ_CONFIG = get_config(
     Namespace(comicbox=Namespace(read=Namespace(formats=("yaml",))))